increase fixture cacheability
